# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Article api",
        "description": "This is our demo api.",
        "termsOfService": "",
        "contact": {
            "email": "rahul@3embed.com"
        },
        "license": {
            "name": "3Embed",
            "url": "https://www.3embed.com"
        },
        "version": "1.0.0"
    },
    "host": "laravel-demo.uberforall.com/api",
    "schemes": [
        "https"
    ],
    "paths": {
        "/articles": {
            "get": {
                "summary": "List all articles",
                "operationId": "getAllArticles",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "summary": "Add New Article",
                "operationId": "postArticle",
                "parameters": [
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "Title.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "formData",
                        "description": "article Body.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/articles/{article}": {
            "get": {
                "summary": "get Article",
                "operationId": "getArticleById",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Target article.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "summary": "update Article",
                "operationId": "putArticle",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Target article.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "Title.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "formData",
                        "description": "article Body",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "summary": "Delete Article",
                "operationId": "deleteArticleById",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Target article.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "summary": "login api",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Target email.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Target password.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "summary": "registration api",
                "operationId": "signup",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Target name.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Target email.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Target password.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "formData",
                        "description": "Target password_confirmation.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
    "definitions": {},
    "securityDefinitions": {
        "default": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "externalDocs": {
        "description": "Find out more about this in our FAQ",
        "url": "https://www.3embed.com"
    }
}